First preliminary working version.
This currently only displays desktop notifications of the snippets of all
messages which are in the inbox. Nothing spectacular and it only displays them
once (the timer is stopped shortly after being triggered for the first time).

The chromium manifest file is generated by a Generator class, as this will not
only help in sharing code (which we currently do not do, for reasons of
laziness), but also for porting the extension to firefox and other browsers.

Another quirk right now is in mailinator.API.TOKEN, which is hardcoded at the
moment. Until we have a nice way to request a token through the users session,
we probably need to scrape the Mailinator Pro settings page.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>